Refactoring top-level make file for building the pochoir executable.
The previous version of the make file was nigh unreadable, unusable,
and unmodifiable. This is being improved by abstracting out the
compiler and associated flags, target and dependencies, and improving
cleanup abilities. This also includes new comments to actully explain
what things are. This version of the make file should be easy to read
and use.