Refactoring top-level run.sh to have cleaner code and conform to better Bash scripting idioms.
I'll be perfectly honest: I have no idea what this script is supposed
to do. It seems like it's supposed to run two executables, loop and
obase, with different size and step values (though there's no
documentation as to what this actually does).

The first change being added is to use proper Bash syntax for numeric
'for' loops. The original used monstrous C-style loops that look
horrid in Bash...because there's actual, easy-to-research syntax for
those types of loops.

The second and more pressing change relates to accessing the
aforementioned loop and obase executables. The new version is adding
path variables for each executable (since they're clearly not at top
level as the old version asserts) and actually checks for the
existence of the executables before trying to call them. The script
also informs the user which of the executables is missing and returns
1 to indicate that it did not finish correctly.